+++ b2sdk/file_version.py
@@ -182,6 +182,13 @@ class FileVersion(BaseFileVersion):
 
         return result
 
+    def get_fresh_state(self) -> 'FileVersion':
+        """
+        Fetch all the information about this file version and return a new FileVersion object.
+        This method does NOT change the object it is called on.
+        """
+        return self.api.get_file_info(self.id_)
+
 
 class DownloadVersion(BaseFileVersion):
     """